requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Authlib==1.3.1
Automat==24.8.1
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.9
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
cachetools==5.3.3
certifi==2024.6.2
cffi @ file:///D:/bld/cffi_1696001730392/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.1.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.7
dataclasses-json==0.6.7
datasets==2.20.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
graphdatascience==1.12
greenlet==3.0.3
grpcio==1.65.1
grpcio-health-checking==1.65.1
grpcio-tools==1.65.1
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hf_transfer==0.1.8
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
hyperlink==21.0.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
importlib_metadata==8.0.0
importlib_resources==6.4.0
incremental==24.7.2
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.2
jedi==0.19.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsondiff==2.2.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
langchain==0.2.12
langchain-community==0.2.6
langchain-core==0.2.28
langchain-openai==0.1.13
langchain-text-splitters==0.2.2
langsmith==0.1.83
llama-cloud==0.0.11
llama-index==0.10.59
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.59
llama-index-embeddings-openai==0.1.11
llama-index-graph-stores-neo4j==0.2.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.32
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
markdown-it-py==3.0.0
MarkupSafe @ file:///D:/bld/markupsafe_1706900062361/work
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.1
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
multidict==6.0.5
multimethod==1.10
multiprocess==0.70.16
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.22.0
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1712540363324/work
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
ollama==0.3.1
openai==1.52.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandoc==2.3
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==4.2.2
plumbum==1.8.3
ply==3.11
preshed==3.0.8
prometheus_client==0.21.0
prompt_toolkit==3.0.48
Protego==0.3.1
protobuf==3.20.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1_modules==0.4.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.3.1
pysbd==0.3.4
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
queuelib==1.7.0
ragas==0.1.11
referencing==0.35.1
regex==2022.10.31
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.7.1
safetensors==0.4.3
Scrapy==2.11.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
service-identity==24.1.0
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacytextblob==0.1.3
SQLAlchemy==2.0.31
stack-data==0.6.3
streamlit==1.35.0
striprtf==0.0.26
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1721004795885/work
tenacity==8.3.0
terminado==0.18.1
textblob==0.15.3
textdistance==4.6.2
tiktoken==0.7.0
tinycss2==1.4.0
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch @ file:///C:/b/abs_8f7uhuge1i/croot/pytorch-select_1717607507421/work
torchvision @ file:///C:/b/abs_61prww4bv9/croot/torchvision_1689079992237/work
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
Twisted==24.7.0
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tyro==0.8.5
tzdata==2024.1
unsloth @ git+https://github.com/unslothai/unsloth.git@4e570be9ae4ced8cdc64e498125708e34942befc
uri-template==1.3.0
urllib3==2.2.1
validators==0.33.0
wasabi==0.10.1
watchdog==4.0.1
wcwidth==0.2.13
weaviate-client==4.6.7
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zope.interface==7.1.0
zstandard==0.23.0
